rewrite code from postgresql

+++ TemperaturyAPI/Models/MinMaxModel.cs
@@ -7,10 +7,10 @@ namespace TemperaturyAPI.Models
     public class MinMaxModel
     {
         public DateOnly DataPomiaru { get; set; }
-        public float MinimalnaTemperaturaGleby { get; set; }
-        public float MaksymalnaTemperaturaGleby { get; set; }
-        public float MinimalnaTemperaturaPowietrza { get; set; }
-        public float MaksymalnaTemperaturaPowietrza { get; set; }
+        public double MinimalnaTemperaturaGleby { get; set; }
+        public double MaksymalnaTemperaturaGleby { get; set; }
+        public double MinimalnaTemperaturaPowietrza { get; set; }
+        public double MaksymalnaTemperaturaPowietrza { get; set; }
     }
 }
